♻️ refactor(graph): modularize graph API into algorithm and utils
Split graph functionality into dedicated modules for better organization.
Extracted graph algorithms (shortest paths, clustering, pagerank) into new Algorithm class,
moved utility methods (neighbor expansion, entity scoring) to Utils class,
and added embedding utilities. Updated references across graph API for cleaner structure.
No functional changes, just improved code maintainability.
